+++ client/app/components/calendar/filters/FilterBox.jsx
