+++ client/app/components/calendar/filters/filters.scss
